refactor: move FeedingContent to commonMain
Splits FeedingContent out of FeedingScreen.kt into commonMain so it can
be exercised by PreviewScreenshotGeneratorTest on the JVM target.
