+++ feature/feeding/build.gradle.kts
@@ -4,11 +4,13 @@ plugins {
 
 kotlin {
     sourceSets {
+        commonMain.dependencies {
+            implementation(project(":core:ui"))
+            implementation(project(":shared"))
+        }
         wasmJsMain.dependencies {
             implementation(project(":core:common"))
             implementation(project(":core:network"))
-            implementation(project(":core:ui"))
-            implementation(project(":shared"))
         }
     }
 }
